requirements.txt
scikit-learn==1.2.1
seaborn
numpy==1.23.5
pandas==1.5.3
ipykernel
yapf
matplotlib
pygeohash
pytest
nltk

google-cloud-bigquery
google-cloud-storage
google-cloud-secret-manager
db-dtypes
pyarrow

joblib
igdb-api-v4

streamlit
requests
datetime
bs4
db_dtypes

flask
